Create M - Full Pyramid
Given an integer N, print a full pyramid of asterisks.
A full pyramid of asterisks of size .N has N lines of asterisks. The first line has 1 asterisk, the second line has 2 asterisks, the third line has 3 asterisks and so on. Each asterisk has a space between them. The asterisks in each line are centered to make it look like a pyramid.
Output: Print the full pyramid of asterisks of the given size.
Do not print any space after the last asterisk in each line.